Frame control is least significant bit first
The frame control diagram is least significant bit first. Adjust to
match on big endian values.

Fix return values of the control frame subtype: they should all return
a 2-tuple with a header and an empty binary.

Add a macro from tcpdump to account for padding. May be useful for the
32/64-bit padding of the wireless tool structs.