Add support for datalink types
Convert wierl_monitor to a gen_server to support DLTs beside the radiotap
format. PRISM headers are still used by some drivers.

Carry the header type in a gen_server, since on Linux the DLT is put in
the hatype field of the struct sockaddr_ll for each packet (for BSDs,
the DLT is retrieved using an ioctl on the fd (BIOCGDLT). We'll need to
support this as well when we port to Mac OS X). This simplifies cleanup
of fd's on exit as well.

Remove close/2,3. Terminating the gen_server only closes the socket,
the mode is not changed. Add mode/1 to change the interface operating
mode. mode/1 is slightly broken, since it needs to lookup up the
interface name in the gen_server state, which will already have been
terminated.