Return the frame checksum
Modify the return value of wierl_frame:type/2 to include the frame
checksum. If the checksum is not available, an empty binary is returned.

The last element of some of the packets is a variable length field
so there is no way of telling if the driver include an FCS. In these
ambiguos cases, the wierl_frame:type/2 returns the atom 'false'. It is
up to the caller to pull out the checksum.

At the moment, the FCS is discarded:

* wierl_monitor: the FCS should be checked and, based upon an option, if
  the check fails, the frame should be discarded

* wierl_monitor: the FCS driver "status" (whether the driver provides
  the FCS) should be kept in the gen_server state. If the FCS status of
  the frame is ambigous, the gen_server will take the status of the last
  frame.

* wierl_monitor: provide an option to override FCS mode