Support unpacking of FCS in ambiguous frames
Keep the driver support for FCS (i.e., whether the driver provides all
packets (including corrupted packets) and the FCS) in the gen_server
state.

Return the checksum to the user as an integer. If the checksum is not
available, return 0 (is 0 a possible checksum? should we return
something else? An atom, -1, ...?).

If the system supports checksums, the gen_server state is updated on
every frame. This is probably inefficient.

The inital state of the fcs field is false. This means it's possible to
have 1 frame incorrectly parsed, e.g., if a frame without a length is
received as the first packet.